active-probe/requirements.txt
dnspython==2.6.1
kafka-python==2.0.2
requests==2.32.3